refactor script.js (separate different functionalities)
