+++ public/add-meal.html
@@ -11,7 +11,7 @@
     <link rel="stylesheet" href="recipe-styles.css">
     <link rel="stylesheet" href="planner-styles.css">
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&display=swap" rel="stylesheet">
-     <script type="module" src="script.js" defer></script>
+     <script type="module" src="script-add-meal.js" defer></script>
 </head>
 <body>
