+++ public/edit-recipe.html
@@ -10,7 +10,7 @@
     <link rel="stylesheet" href="home-styles.css">
     <link rel="stylesheet" href="profile-styles.css">
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600&display=swap" rel="stylesheet">
-     <script type="module" src="script.js"></script>
+     <script type="module" src="script-edit-recipe.js"></script>
 </head>
 
 <body>
