+++ public/home-page.html
@@ -9,7 +9,6 @@
     <link rel="stylesheet" href="user-profile-styles.css">
     <link rel="stylesheet" href="home-styles.css">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600&display=swap">
-    <script type="module" src="script.js"></script>
     <script type="module" src="script-homepage.js" defer></script>
 </head>
