+++ public/meal-planner.html
@@ -10,7 +10,7 @@
         <link rel="stylesheet" href="profile-styles.css">
         <link rel="stylesheet" href="planner-styles.css">
         <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&display=swap" rel="stylesheet">
-         <script type="module" src="script.js" defer></script>
+        <script type="module" src="script-meal-planner.js" defer></script>
 </head>
 <body>
