+++ public/script.js
@@ -6,15 +6,6 @@
 // CONFIGURATION
 const API_BASE_URL = 'http://localhost:4002/api';
 const TOKEN_KEY = 'mealmajor_token';
-const MAX_SIZE_MB = 50; // max video upload size
-
-// Global variable to store all recipes for filtering
-let allRecipes = [];
-
-// Variables for undo delete meal plan
-let lastDeletedMeal = null;
-let undoTimeout = null;
-
 
 // UTILITY FUNCTIONS
 /**
@@ -23,65 +14,48 @@ let undoTimeout = null;
 export function saveToken(token) {
   localStorage.setItem(TOKEN_KEY, token);
 }
-
 /**
  * Retrieve authentication token from localStorage
  */
 export function getToken() {
   return localStorage.getItem(TOKEN_KEY);
 }
-
 /**
  * Remove authentication token (logout)
  */
 export function clearToken() {
   localStorage.removeItem(TOKEN_KEY);
 }
-
 /**
  * Check if user is authenticated
  */
 export function isAuthenticated() {
   return !!getToken();
 }
-
 /**
  * Display error message in a form field
  */
-function showError(elementId, message) {
+export function showError(elementId, message) {
   const errorElement = document.getElementById(elementId);
   if (errorElement) {
     errorElement.textContent = message;
     errorElement.style.display = 'block';
   }
 }
-
-/**
- * Clear error message from a form field
- */
-function clearError(elementId) {
-  const errorElement = document.getElementById(elementId);
-  if (errorElement) {
-    errorElement.textContent = '';
-    errorElement.style.display = 'none';
-  }
-}
-
 /**
  * Clear all error messages on a page
  */
-function clearAllErrors() {
+export function clearAllErrors() {
   const errorElements = document.querySelectorAll('.error');
   errorElements.forEach(el => {
     el.textContent = '';
     el.style.display = 'none';
   });
 }
-
 /**
  * Show loading state on button
  */
-function setButtonLoading(button, isLoading) {
+export function setButtonLoading(button, isLoading) {
   if (isLoading) {
     button.disabled = true;
     button.dataset.originalText = button.querySelector('span')?.textContent || button.textContent;
@@ -101,12 +75,14 @@ function setButtonLoading(button, isLoading) {
     }
   }
 }
-
+/**
+ * Get initials of the user for avatar display in navbar
+ */
 export async function getInitials() {
   // If not authenticated, skip
   if (!isAuthenticated()) return;
   
-  const response = await fetch(`/api/profile`, {
+  const response = await fetch(`${API_BASE_URL}/profile`, {
       method: 'GET',
       headers: {
         'Authorization': `Bearer ${getToken()}`,
@@ -116,88 +92,30 @@ export async function getInitials() {
   const user = await response.json();
   const avatar = document.getElementById("avatarNav");
   if (avatar && user.firstName && user.lastName) {
-  avatar.textContent = user.firstName[0] + user.lastName[0];
+    avatar.textContent = user.firstName[0] + user.lastName[0];
+    return avatar.textContent;
   }
 }
-
-// VALIDATION FUNCTIONS
 /**
  * Validate email format
  */
-function validateEmail(email) {
+export function validateEmail(email) {
   const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
   return emailRegex.test(email);
 }
-
 /**
- * Validate password strength
- * Requirements: At least 8 characters, 1 number, 1 special character
+ * logout function to clear token and redirect to login page
  */
-function validatePassword(password) {
-  if (password.length < 8) {
-    return { valid: false, message: 'Password must be at least 8 characters long' };
-  }
-  if (!/\d/.test(password)) {
-    return { valid: false, message: 'Password must contain at least one number' };
-  }
-  if (!/[!@#$%^&*(),.?":{}|<>]/.test(password)) {
-    return { valid: false, message: 'Password must contain at least one special character' };
-  }
-  return { valid: true, message: '' };
-}
-
-
-// API FUNCTIONS
-/**
- * Login user
- */
-async function loginUser(email, password) {
-  try {
-    const response = await fetch(`${API_BASE_URL}/auth/login`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ email, password }),
+export function logout() {
+  const logoutLink = document.querySelector('.nav-link.logout');
+    if (logoutLink) {
+      logoutLink.addEventListener('click', (e) => {
+        e.preventDefault();
+        clearToken();
+        window.location.href = 'login-page.html';
     });
-
-    const data = await response.json();
-
-    if (!response.ok) {
-      throw new Error(data.message || 'Login failed');
-    }
-
-    return { success: true, data };
-  } catch (error) {
-    return { success: false, error: error.message };
   }
 }
-
-/**
- * Register new user
- */
-async function registerUser(firstName, lastName, email, password) {
-  try {
-    const response = await fetch(`${API_BASE_URL}/auth/register`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ firstName, lastName, email, password }),
-    });
-
-    const data = await response.json();
-
-    if (!response.ok) {
-      throw new Error(data.message || 'Registration failed');
-    }
-
-    return { success: true, data };
-  } catch (error) {
-    return { success: false, error: error.message };
-  }
-}
-
 /**
  * Get user profile
  */
@@ -227,96 +145,6 @@ export async function getUserProfile() {
     return { success: false, error: error.message };
   }
 }
-
-/**
- * Update user profile
- */
-async function updateUserProfile(updates) {
-  try {
-    const token = getToken();
-    if (!token) {
-      throw new Error('No authentication token found');
-    }
-
-    const response = await fetch(`${API_BASE_URL}/profile`, {
-      method: 'PUT',
-      headers: {
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(updates),
-    });
-
-    const data = await response.json();
-
-    if (!response.ok) {
-      throw new Error(data.message || 'Failed to update profile');
-    }
-
-    return { success: true, data };
-  } catch (error) {
-    return { success: false, error: error.message };
-  }
-}
-
-/**
- * Fetch recipes from the API
- */
-async function fetchRecipes(ownerId = null) {
-  try {
-    let url = `${API_BASE_URL}/recipes`;
-    if (ownerId) {
-      url += `?ownerId=${ownerId}`;
-    }
-
-    const response = await fetch(url, {
-      headers: {
-        'Authorization': `Bearer ${getToken()}`
-      }
-    });
-
-    if (!response.ok) {
-      throw new Error('Failed to fetch recipes');
-    }
-
-    const data = await response.json();
-    return { success: true, data };
-  } catch (error) {
-    console.error('Error fetching recipes:', error);
-    return { success: false, error: error.message };
-  }
-}
-
-// MEAL PLAN FUNCTIONS
-async function createMealPlanForWeek(weekStartDate, weekEndDate) {
-  try {
-    const response = await fetch(`${API_BASE_URL}/mealPlan`, {
-      method: 'POST',
-      headers: {
-        'Authorization': `Bearer ${getToken()}`,
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        week_start_date: weekStartDate,
-        week_end_date: weekEndDate,
-        name: 'Weekly Meal Plan'
-      })
-    });
-
-    const data = await response.json();
-
-    if (!response.ok) {
-      throw new Error(data.message || 'Failed to create meal plan');
-    }
-
-    return { success: true, data };
-  } catch (error) {
-    return { success: false, error: error.message };
-  }
-}
-
-
-// PROFILE FUNCTIONS
 /**
  * Check profile completion status
  */
@@ -386,10 +214,51 @@ export async function checkProfileCompletion() {
     });
   }
 }
+/**
+ * fetch ratings for a recipe
+ */
+export async function fetchRecipeRatings(recipeId) {
+  try {
+    const response = await fetch(`${API_BASE_URL}/recipeRatings/${recipeId}`, {
+      headers: {
+        'Authorization': `Bearer ${getToken()}`
+      }
+    });
 
+    if (!response.ok) return [];
+    return await response.json();
+  } catch (error) {
+    console.error("Error fetching recipe ratings:", error);
+    return [];
+  }
+}
+/**
+ * Fetch average rating for a recipe
+ */
+export async function fetchAverageRating(recipeId) {
+  try {
+    const response = await fetch(`${API_BASE_URL}/averageRating/${recipeId}`, {
+      headers: {
+        'Authorization': `Bearer ${getToken()}`
+      }
+    });
+
+    if (!response.ok) return 0;
+    const avg = await response.json();
+    return Number(avg) || 0;
+  } catch (error) {
+    console.error("Error fetching average rating:", error);
+    return 0;
+  }
+}
+
+export function renderStars(avgRating) {
+  const rounded = Math.round(avgRating);
+  return '★'.repeat(rounded) + '☆'.repeat(5 - rounded);
+}
 /**
  * Show profile prompt
- */
+*/
 function showProfilePrompt(missingFields) {
   const prompt = document.getElementById('profilePrompt');
   if (prompt) {
@@ -411,1397 +280,10 @@ function showProfilePrompt(missingFields) {
     }
   }
 }
-
 /**
- * Show success toast
+ * Function to create a new meal plan item
  */
-function showSuccessToast(message) {
-  const toast = document.getElementById('successToast');
-  const toastMessage = document.querySelector('.toast-message');
-
-  if (toast && toastMessage) {
-    toastMessage.textContent = message || 'Profile updated successfully!';
-    toast.classList.remove('hidden');
-
-    // Auto hide after 3 seconds
-    setTimeout(() => {
-      toast.classList.add('hidden');
-    }, 3000);
-  }
-}
-
-/**
- * Load user profile and populate edit form
- */
-async function loadUserProfileForEdit() {
-  const result = await getUserProfile();
-
-  if (result.success) {
-    const profile = result.data;
-    console.log('Profile loaded:', profile);
-
-    // Populate basic fields
-    if (document.getElementById('firstName')) {
-      document.getElementById('firstName').value = profile.firstName || '';
-    }
-    if (document.getElementById('lastName')) {
-      document.getElementById('lastName').value = profile.lastName || '';
-    }
-    if (document.getElementById('email')) {
-      document.getElementById('email').value = profile.email || '';
-    }
-
-    // Populate health metrics
-    if (document.getElementById('editAge')) {
-      document.getElementById('editAge').value = profile.age || '';
-    }
-    if (document.getElementById('editWeight')) {
-      document.getElementById('editWeight').value = profile.weight || '';
-    }
-    if (document.getElementById('editHeight')) {
-      document.getElementById('editHeight').value = profile.height || '';
-    }
-
-    // Populate diet preferences
-    if (profile.dietPreferences && Array.isArray(profile.dietPreferences)) {
-      profile.dietPreferences.forEach(diet => {
-        const checkbox = document.querySelector(`input[name="diet"][value="${diet}"]`);
-        if (checkbox) checkbox.checked = true;
-      });
-    }
-
-    // Populate allergies
-    if (profile.allergies && Array.isArray(profile.allergies)) {
-      const customAllergies = [];
-      profile.allergies.forEach(allergy => {
-        const checkbox = document.querySelector(`input[name="allergy"][value="${allergy}"]`);
-        if (checkbox) {
-          checkbox.checked = true;
-        } else {
-          customAllergies.push(allergy);
-        }
-      });
-      //extra allergies that are not predefined
-      if (customAllergies.length > 0) {
-        const otherAllergiesArea = document.getElementById('otherAllergies');
-        if (otherAllergiesArea) {
-          otherAllergiesArea.value = customAllergies.join(', ');
-        }
-      }
-    }
-
-    // Populate cooking skill
-    if (document.getElementById('cookingSkill')) {
-      document.getElementById('cookingSkill').value = profile.cookingSkill || '';
-    }
-    // Populate meal prep time
-    if (document.getElementById('mealPrepTime')) {
-      document.getElementById('mealPrepTime').value = profile.mealPrepTime || '';
-    }
-
-    // Populate budget range
-    if (document.getElementById('budgetRange')) {
-      document.getElementById('budgetRange').value = profile.budgetRange || '';
-    }
-  } else {
-    console.error('Failed to load profile:', result.error);
-    // If token is invalid, redirect to login
-    if (result.error.includes('token') || result.error.includes('auth')) {
-      clearToken();
-      window.location.href = 'login-page.html';
-    }
-  }
-}
-
-
-// RECIPE FUNCTIONS
-
-/**
- * Load all recipes for filtering
- */
-async function loadAllRecipes() {
-  const result = await fetchRecipes();
-  if (result.success) {
-    allRecipes = result.data;
-    console.log('Loaded all recipes for filtering:', allRecipes.length);
-  }
-}
-
-/**
- * Load My Recipes
- */
-async function loadMyRecipes() {
-  console.log('Loading My Recipes...');
-
-  const myRecipesGrid = document.getElementById('myRecipesGrid');
-  const emptyState = document.getElementById('myRecipesEmpty');
-
-  if (!myRecipesGrid) return;
-
-  try {
-    // Get current user profile to get their ID
-    const profileResult = await getUserProfile();
-
-    if (!profileResult.success) {
-      console.error('Failed to get user profile:', profileResult.error);
-      return;
-    }
-
-    const ownerId = profileResult.data.id;
-
-    // Fetch recipes owned by this user
-    const result = await fetchRecipes(ownerId);
-
-    if (result.success) {
-      const recipes = result.data;
-      console.log('My Recipes fetched:', recipes.length);
-
-      if (recipes.length === 0) {
-        // Show empty state
-        if (emptyState) emptyState.style.display = 'flex';
-        myRecipesGrid.innerHTML = '';
-      } else {
-        // Hide empty state
-        if (emptyState) emptyState.style.display = 'none';
-
-        // Display recipes
-        myRecipesGrid.innerHTML = '';
-        recipes.forEach(recipe => {
-          const card = createRecipeCard(recipe, true);
-          myRecipesGrid.appendChild(card);
-        });
-      }
-    }
-  } catch (error) {
-    console.error('Error in loadMyRecipes:', error);
-  }
-}
-
-// Load Explore Recipes (all recipes with pagination)
-async function loadExploreRecipes() {
-  console.log("🌍 Loading explore recipes...");
-
-  const grid = document.getElementById("generalRecipesGrid");
-  if (!grid) return;
-
-  try {
-    const response = await fetch(`${API_BASE_URL}/recipes/explore`, {
-      headers: {
-        Authorization: `Bearer ${getToken()}`
-      }
-    });
-
-    if (!response.ok) throw new Error("Failed to fetch explore recipes");
-
-    const recipes = await response.json();
-
-    console.log("Explore recipes:", recipes.length);
-
-    allRecipes = recipes;
-    grid.innerHTML = "";
-
-    for (const recipe of recipes) {
-      const card = await createPublicRecipeCard(recipe);
-      grid.appendChild(card);
-    }
-
-  } catch (error) {
-    console.error("Error loading explore recipes:", error);
-  }
-}
-
-/**
- * Load General Recipes
- */
-// async function loadGeneralRecipes() {
-
-//   const generalGrid = document.getElementById('generalRecipesGrid');
-//   if (!generalGrid) return;
-
-//   try {
-//     // Fetch all recipes
-//     const result = await fetchRecipes();
-
-//     if (result.success) {
-//       // Filter to ONLY show recipes with ownerId = 1 (seed recipes)
-//       let recipes = result.data.filter(recipe => recipe.ownerId === 1);
-
-//       // Assign emojis based on recipe names
-//       recipes = recipes.map(recipe => {
-//         // Create a copy of the recipe with an emoji
-//         const recipeWithEmoji = { ...recipe };
-
-//         // Assign emoji based on name
-//         if (recipe.name.includes("Chickpea")) recipeWithEmoji.emoji = "🌯";
-//         else if (recipe.name.includes("Shrimp")) recipeWithEmoji.emoji = "🍝";
-//         else if (recipe.name.includes("Fried Rice")) recipeWithEmoji.emoji = "🍚";
-//         else if (recipe.name.includes("Beef Chili")) recipeWithEmoji.emoji = "🥘";
-//         else if (recipe.name.includes("Salmon")) recipeWithEmoji.emoji = "🐟";
-//         else if (recipe.name.includes("Shawarma")) recipeWithEmoji.emoji = "🥙";
-//         else recipeWithEmoji.emoji = "🍽️"; // Default emoji
-
-//         return recipeWithEmoji;
-//       });
-
-//       console.log('General recipes fetched (ownerId=1):', recipes.length);
-
-//       // Store in global variable for filtering
-//       allRecipes = recipes;
-
-//       // Clear and populate grid
-//       generalGrid.innerHTML = '';
-
-//       recipes.forEach(recipe => {
-//         const card = createRecipeCard(recipe, false);
-//         generalGrid.appendChild(card);
-//       });
-//     }
-//   } catch (error) {
-//     console.error('Error in loadGeneralRecipes:', error);
-//   }
-// }
-
-/**
- * Filter recipes by search term
- */
-async function filterRecipesBySearch(searchTerm) {
-  console.log('Searching for:', searchTerm);
-
-  const generalGrid = document.querySelector('.section:last-child .recipe-cards-grid');
-  if (!generalGrid) return;
-
-  // If search is empty, show all recipes
-  if (!searchTerm || searchTerm === '') {
-    await displayFilteredRecipes(allRecipes);
-    return;
-  }
-
-  // Convert search term to lowercase for case-insensitive comparison
-  const term = searchTerm.toLowerCase();
-
-  // Filter recipes by name (partial match)
-  const filtered = allRecipes.filter(recipe => {
-    return recipe.name && recipe.name.toLowerCase().includes(term);
-  });
-
-  console.log('Search results:', filtered.length);
-
-  if (filtered.length === 0) {
-    generalGrid.innerHTML = '<p class="no-results">No recipes match your search</p>';
-  } else {
-    await displayFilteredRecipes(filtered);
-  }
-}
-
-/**
- * Filter recipes based on selected filters
- */
-async function filterRecipes() {
-  console.log('Filtering recipes...');
-
-  // Get all selected filter values by name attribute (more reliable)
-  const timeChecked = Array.from(document.querySelectorAll('input[name="time"]:checked')).map(cb => cb.value);
-  const difficultyChecked = Array.from(document.querySelectorAll('input[name="difficulty"]:checked')).map(cb => cb.value);
-  const costChecked = Array.from(document.querySelectorAll('input[name="cost"]:checked')).map(cb => cb.value);
-  const dietaryChecked = Array.from(document.querySelectorAll('input[name="diet"]:checked')).map(cb => cb.value);
-  const allergenChecked = Array.from(document.querySelectorAll('input[name="allergy"]:checked')).map(cb => cb.value);
-
-  console.log('Selected filters:', {
-    time: timeChecked,
-    difficulty: difficultyChecked,
-    cost: costChecked,
-    dietary: dietaryChecked,
-    allergens: allergenChecked
-  });
-
-  // If no filters selected, show all recipes
-  if (timeChecked.length === 0 && difficultyChecked.length === 0 &&
-    costChecked.length === 0 && dietaryChecked.length === 0 &&
-    allergenChecked.length === 0) {
-    await displayFilteredRecipes(allRecipes);
-    return;
-  }
-
-  // Filter recipes
-  const filtered = allRecipes.filter(recipe => {
-    // TIME FILTER
-    if (timeChecked.length > 0) {
-      const timeMatch = timeChecked.some(time => {
-        if (time === 'under-15' && recipe.prep_time < 15) return true;
-        if (time === '15-30' && recipe.prep_time >= 15 && recipe.prep_time <= 30) return true;
-        if (time === '30-60' && recipe.prep_time >= 30 && recipe.prep_time <= 60) return true;
-        if (time === '60plus' && recipe.prep_time > 60) return true;
-        return false;
-      });
-      if (!timeMatch) return false;
-    }
-
-    // DIFFICULTY FILTER (case insensitive)
-    if (difficultyChecked.length > 0) {
-      const difficultyMatch = difficultyChecked.some(diff =>
-        recipe.difficulty && recipe.difficulty.toLowerCase() === diff.toLowerCase()
-      );
-      if (!difficultyMatch) return false;
-    }
-
-    // COST FILTER (case insensitive)
-    if (costChecked.length > 0) {
-      const costMatch = costChecked.some(cost =>
-        recipe.cost && recipe.cost.toLowerCase() === cost.toLowerCase()
-      );
-      if (!costMatch) return false;
-    }
-
-    // DIETARY FILTER (case insensitive)
-    if (dietaryChecked.length > 0 && recipe.dietary_tags) {
-      const dietaryMatch = dietaryChecked.some(diet => {
-        // Use case-insensitive comparison for all dietary tags
-        return recipe.dietary_tags.some(tag =>
-          tag.toLowerCase() === diet.toLowerCase()
-        );
-      });
-      if (!dietaryMatch) return false;
-    }
-
-    // ALLERGEN FILTER - Hide recipes with selected allergens (case insensitive)
-    if (allergenChecked.length > 0 && recipe.allergens) {
-      const hasAllergen = allergenChecked.some(allergen => {
-        return recipe.allergens.some(recipeAllergen =>
-          recipeAllergen.toLowerCase() === allergen.toLowerCase()
-        );
-      });
-      if (hasAllergen) return false; // Hide recipes with these allergens
-    }
-
-    return true;
-  });
-
-  console.log('Filtered recipes:', filtered.length);
-  await displayFilteredRecipes(filtered);
-}
-/**
- * Display filtered recipes in the general grid
- */
-async function displayFilteredRecipes(recipes) {
-  const generalGrid = document.querySelector('.section:last-child .recipe-cards-grid');
-  if (!generalGrid) return;
-
-  generalGrid.innerHTML = '';
-
-  if (recipes.length === 0) {
-    generalGrid.innerHTML = '<p class="no-results">No recipes match your filters</p>';
-    return;
-  }
-
-  for (const recipe of recipes) {
-    const card = await createPublicRecipeCard(recipe);
-    generalGrid.appendChild(card);
-  }
-}
-
-async function fetchRecipeRatings(recipeId) {
-  try {
-    const response = await fetch(`${API_BASE_URL}/recipeRatings/${recipeId}`, {
-      headers: {
-        'Authorization': `Bearer ${getToken()}`
-      }
-    });
-
-    if (!response.ok) return [];
-    return await response.json();
-  } catch (error) {
-    console.error("Error fetching recipe ratings:", error);
-    return [];
-  }
-}
-
-async function fetchAverageRating(recipeId) {
-  try {
-    const response = await fetch(`${API_BASE_URL}/averageRating/${recipeId}`, {
-      headers: {
-        'Authorization': `Bearer ${getToken()}`
-      }
-    });
-
-    if (!response.ok) return 0;
-    const avg = await response.json();
-    return Number(avg) || 0;
-  } catch (error) {
-    console.error("Error fetching average rating:", error);
-    return 0;
-  }
-}
-
-// Fetch recipe video - handles both URL and file responses
-/* async function fetchRecipeVideo(recipeId) {
-  try {
-    const response = await fetch(`${API_BASE_URL}/recipes/${recipeId}/video`, {
-      headers: {
-        'Authorization': `Bearer ${getToken()}`
-      }
-    });
-
-    if (!response.ok) return null;
-
-    const contentType = response.headers.get("content-type") || "";
-
-    if (contentType.includes("application/json")) {
-      const data = await response.json();
-      return data?.videoUrl ? { type: "url", value: data.videoUrl } : null;
-    }
-
-    if (contentType.startsWith("video")) {
-      const blob = await response.blob();
-      return {
-        type: "file",
-        value: URL.createObjectURL(blob),
-        mimeType: contentType
-      };
-    }
-
-    return null;
-  } catch (error) {
-    console.error("Error fetching recipe video:", error);
-    return null;
-  }
-} */
-
-function renderStars(avgRating) {
-  const rounded = Math.round(avgRating);
-  return '★'.repeat(rounded) + '☆'.repeat(5 - rounded);
-}
-
-function getUserInitialsFromEmail(email = "") {
-  const name = email.split("@")[0] || "?";
-  return name.substring(0, 2).toUpperCase();
-}
-
-// Submit recipe rating and comment
-async function submitRecipeRatingAndComment(recipeId, rating, comment) {
-  try {
-    const response = await fetch(`${API_BASE_URL}/recipeRatings`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${getToken()}`
-      },
-      body: JSON.stringify({
-        recipeId,
-        rating,
-        comment
-      })
-    });
-
-    const data = await response.json();
-
-    if (!response.ok) {
-      throw new Error(data.message || "Failed to submit rating/comment");
-    }
-
-    return { success: true, data };
-  } catch (error) {
-    console.error("Error submitting rating/comment:", error);
-    return { success: false, error: error.message };
-  }
-}
-
-/**
- * Create recipe card element
- */
-function createRecipeCard(recipe, isMyRecipe = false) {
-  const details = document.createElement('details');
-  details.className = 'recipe-card-full';
-
-  // Determine difficulty class
-  let difficultyClass = 'tag-easy';
-  if ((recipe.difficulty || '').toLowerCase() === 'medium') difficultyClass = 'tag-medium';
-  if ((recipe.difficulty || '').toLowerCase() === 'hard') difficultyClass = 'tag-hard';
-
-  // Cost display
-  let costDisplay = '💰 Low';
-  if ((recipe.cost || '').toLowerCase() === 'medium') costDisplay = '💰💰 Medium';
-  if ((recipe.cost || '').toLowerCase() === 'high') costDisplay = '💰💰💰 High';
-
-  // Dietary tags display
-  const dietaryDisplay = recipe.dietary_tags && recipe.dietary_tags.length > 0
-    ? recipe.dietary_tags.join(' · ')
-    : '';
-
-  // Summary
-  const summary = document.createElement('summary');
-  summary.className = 'recipe-card-top';
-  summary.innerHTML = `
-<div class="recipe-card-emoji">${recipe.emoji || '🍽️'}</div>
-    <div class="recipe-card-summary">
-      <div class="recipe-card-tags">
-        <span class="tag ${difficultyClass}">${recipe.difficulty || 'Easy'}</span>
-        <span class="tag tag-cost">${costDisplay}</span>
-      </div>
-      <h3 class="recipe-card-title">${recipe.name}</h3>
-      <div class="recipe-card-meta">
-        <span>⏱️ ${recipe.prep_time} min</span>
-        <span>${dietaryDisplay}</span>
-      </div>
-    </div>
-    <span class="recipe-card-chevron">▼</span>
-  `;
-
-  // Details content
-  const detailsDiv = document.createElement('div');
-  detailsDiv.className = 'recipe-card-details';
-
-  // Format ingredients as list items
-  const ingredientsList = recipe.ingredients && Array.isArray(recipe.ingredients)
-    ? recipe.ingredients.map(ingredient => `<li>${ingredient}</li>`).join('')
-    : '<li>No ingredients listed</li>';
-
-  // Format steps as list items
-  const stepsList = recipe.prep_steps && Array.isArray(recipe.prep_steps)
-    ? recipe.prep_steps.map(step => `<li>${step}</li>`).join('')
-    : '<li>No steps provided</li>';
-
-  // Action buttons based on recipe ownership
-  let actionButtons = '';
-  if (isMyRecipe) {
-    actionButtons = `
-      <div class="recipe-card-actions">
-        <button class="btn-edit-recipe" onclick="editRecipe(${recipe.id})">Edit</button>
-        <button class="btn-delete-recipe" onclick="deleteRecipe(${recipe.id})">Delete</button>
-      </div>
-    `;
-  } else {
-    actionButtons = `
-      <button class="btn-save-recipe" onclick="saveRecipeToMyCollection(${recipe.id})">+ Save to My Recipes</button>
-    `;
-  }
-
-  detailsDiv.innerHTML = `
-    <div class="recipe-card-ingredients">
-      <strong>Ingredients:</strong>
-<ul>
-  ${ingredientsList}
-</ul>
-    </div>
-    <div class="recipe-card-steps">
-      <strong>Steps:</strong>
-      <ol>
-        ${stepsList}
-      </ol>
-    </div>
-    ${actionButtons}
-  `;
-
-  details.appendChild(summary);
-  details.appendChild(detailsDiv);
-
-  return details;
-}
-
-async function createPublicRecipeCard(recipe) {
-  const details = document.createElement('details');
-  details.className = 'recipe-card-full public-recipe-card';
-  details.dataset.recipeId = recipe.id;
-
-  const avgRating = await fetchAverageRating(recipe.id);
-  const ratings = await fetchRecipeRatings(recipe.id);
-  const comments = ratings.filter(r => r.comment && r.comment.trim() !== "");
-  //const video = await fetchRecipeVideo(recipe.id);
-
-  let difficultyClass = 'tag-easy';
-  const difficultyValue = (recipe.difficulty || 'Easy');
-  if (difficultyValue.toLowerCase() === 'medium') difficultyClass = 'tag-medium';
-  if (difficultyValue.toLowerCase() === 'hard') difficultyClass = 'tag-hard';
-
-  let costDisplay = '💰 Low';
-  if ((recipe.cost || '').toLowerCase() === 'medium') costDisplay = '💰💰 Medium';
-  if ((recipe.cost || '').toLowerCase() === 'high') costDisplay = '💰💰💰 High';
-
-  const dietaryDisplay = recipe.dietary_tags && recipe.dietary_tags.length > 0
-    ? recipe.dietary_tags.join(' · ')
-    : '';
-
-  const ownerName = recipe.owner?.email
-    ? `@${recipe.owner.email.split('@')[0]}`
-    : '@user';
-
-  const ownerInitials = recipe.owner
-    ? `${(recipe.owner.firstName || '?')[0]}${(recipe.owner.lastName || '?')[0]}`
-    : '??';
-
-  const firstVideo = Array.isArray(recipe.videos) && recipe.videos.length > 0
-  ? recipe.videos[0]
-  : null;
-
-  const videoLink = firstVideo?.videoUrl || null;
-
-  const videoHTML = videoLink
-    ? `
-      <a href="${videoLink}" target="_blank" class="video-link">
-        ▶ Watch Video
-      </a>
-    `
-    : `
-      <p>No video available.</p>
-    `;
-
-  const commentsHTML = comments.length
-    ? comments.map(comment => `
-        <div class="comment-item">
-          <div class="comment-avatar">${getUserInitialsFromEmail(comment.user?.email || "")}</div>
-          <div class="comment-body">
-            <div class="comment-header">
-              <span class="comment-username">@${comment.user?.email ? comment.user.email.split('@')[0] : 'user'}</span>
-              <span class="comment-time">${new Date(comment.updated_at || comment.created_at).toLocaleDateString()}</span>
-            </div>
-            <p class="comment-text">${comment.comment}</p>
-          </div>
-        </div>
-      `).join('')
-    : '<p>No comments yet.</p>';
-
-  details.innerHTML = `
-    <summary class="recipe-card-top">
-      <span class="recipe-card-emoji">${recipe.emoji || '🍽️'}</span>
-      <div class="recipe-card-summary">
-        <div class="recipe-card-tags">
-          <span class="tag ${difficultyClass}">${difficultyValue}</span>
-          <span class="tag tag-cost">${costDisplay}</span>
-          ${dietaryDisplay ? `<span class="tag tag-diet">${dietaryDisplay}</span>` : ''}
-        </div>
-        <h3 class="recipe-card-title">${recipe.name}</h3>
-        <div class="recipe-card-meta">
-          <span>⏱️ ${recipe.prep_time ?? '-'} min</span>
-        </div>
-        <div class="recipe-card-author-row">
-          <div class="recipe-author-avatar">${ownerInitials}</div>
-          <span class="recipe-author-name">${ownerName}</span>
-          <div class="recipe-star-rating">${renderStars(avgRating)}</div>
-          <span class="recipe-rating-count">(${ratings.length})</span>
-        </div>
-      </div>
-      <span class="recipe-card-chevron">▼</span>
-    </summary>
-
-    <div class="recipe-card-details">
-      <div class="recipe-video">
-        <h4>🎥 Video</h4>
-        ${videoHTML}
-      </div>
-
-      <div class="recipe-card-ingredients">
-        <strong>🛒 Ingredients</strong><br>
-        ${(recipe.ingredients || []).join(' · ')}
-      </div>
-
-      <div class="recipe-card-steps">
-        <strong>📋 Steps</strong>
-        <ol>
-          ${(recipe.prep_steps || []).map(step => `<li>${step}</li>`).join('')}
-        </ol>
-      </div>
-
-      <div class="recipe-comments">
-        <h4 class="comments-heading">💬 Comments <span class="comments-count">${comments.length}</span></h4>
-        <div class="comments-list">${commentsHTML}</div>
-      </div>
-
-      <div class="recipe-rate-comment">
-        <h4>⭐ Rate & 💬 Comment</h4>
-        <div class="stars-input">
-          <input type="radio" id="star5-${recipe.id}" name="rating-${recipe.id}" value="5"><label for="star5-${recipe.id}">★</label>
-          <input type="radio" id="star4-${recipe.id}" name="rating-${recipe.id}" value="4"><label for="star4-${recipe.id}">★</label>
-          <input type="radio" id="star3-${recipe.id}" name="rating-${recipe.id}" value="3"><label for="star3-${recipe.id}">★</label>
-          <input type="radio" id="star2-${recipe.id}" name="rating-${recipe.id}" value="2"><label for="star2-${recipe.id}">★</label>
-          <input type="radio" id="star1-${recipe.id}" name="rating-${recipe.id}" value="1"><label for="star1-${recipe.id}">★</label>
-        </div>
-        <textarea class="comment-input" id="comment-${recipe.id}" placeholder="Add a comment..." rows="2"></textarea>
-        <button class="btn-submit-comment" data-recipe-id="${recipe.id}">Submit</button>
-      </div>
-
-      <button class="btn-save-recipe" data-save-recipe-id="${recipe.id}">+ Save to My Recipes</button>
-    </div>
-  `;
-
-  const submitBtn = details.querySelector('.btn-submit-comment');
-  submitBtn.addEventListener('click', async (e) => {
-    e.preventDefault();
-
-    const selectedRating = details.querySelector(`input[name="rating-${recipe.id}"]:checked`);
-    const commentText = details.querySelector(`#comment-${recipe.id}`).value.trim();
-
-    if (!selectedRating) {
-      alert("Please select a rating.");
-      return;
-    }
-
-    const result = await submitRecipeRatingAndComment(
-      recipe.id,
-      selectedRating ? parseInt(selectedRating.value, 10) : null,
-      commentText
-    );
-
-    if (result.success) {
-      alert("Comment/rating submitted!");
-      await loadExploreRecipes();
-    } else {
-      alert(result.error || "Failed to submit.");
-    }
-  });
-
-  const saveBtn = details.querySelector('[data-save-recipe-id]');
-  saveBtn.addEventListener('click', async (e) => {
-    e.preventDefault();
-    await saveRecipeToMyCollection(recipe.id);
-  });
-
-  return details;
-}
-
-// Functions for edit/delete/save
-async function editRecipe(id) {
-  console.log('Edit recipe:', id);
-  window.location.href = `edit-recipe.html?id=${id}`;
-}
-
-async function deleteRecipe(id) {
-  if (confirm('Are you sure you want to delete this recipe?')) {
-    try {
-      const response = await fetch(`${API_BASE_URL}/recipes/${id}`, {
-        method: 'DELETE',
-        headers: {
-          'Authorization': `Bearer ${getToken()}`
-        }
-      });
-
-      if (response.ok) {
-        alert('Recipe deleted successfully!');
-        // Reload recipes
-        loadMyRecipes();
-        loadAllRecipes();
-        //loadGeneralRecipes();
-      } else {
-        alert('Failed to delete recipe');
-      }
-    } catch (error) {
-      console.error('Error deleting recipe:', error);
-      alert('Error deleting recipe');
-    }
-  }
-}
-
-// Save a recipe from explore to My Recipes by copying its data and creating a new recipe owned by the user
-async function saveRecipeToMyCollection(recipeId) {
-  try {
-    const response = await fetch(`${API_BASE_URL}/recipes/${recipeId}`, {
-      headers: {
-        'Authorization': `Bearer ${getToken()}`
-      }
-    });
-
-    if (!response.ok) {
-      throw new Error("Failed to load recipe");
-    }
-
-    const recipe = await response.json();
-
-    const recipeData = {
-      name: recipe.name,
-      ingredients: recipe.ingredients || [],
-      prep_time: recipe.prep_time,
-      prep_steps: recipe.prep_steps || [],
-      cost: recipe.cost,
-      difficulty: recipe.difficulty,
-      dietary_tags: recipe.dietary_tags || [],
-      allergens: recipe.allergens || []
-    };
-
-    const createResponse = await fetch(`${API_BASE_URL}/recipes`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${getToken()}`
-      },
-      body: JSON.stringify(recipeData)
-    });
-
-    const createData = await createResponse.json();
-
-    if (!createResponse.ok) {
-      throw new Error(createData.message || "Failed to save recipe");
-    }
-
-    alert("Recipe saved to My Recipes!");
-    loadMyRecipes();
-  } catch (error) {
-    console.error("Error saving recipe to collection:", error);
-    alert(error.message || "Failed to save recipe");
-  }
-}
-
-
-// CREATE/EDIT RECIPE FUNCTIONS
-
-/**
- * Helper functions to get selected tags
- */
-function getSelectedDietaryTags() {
-  const checkboxes = document.querySelectorAll('input[name="dietary"]:checked');
-  return Array.from(checkboxes).map(cb => cb.value);
-}
-
-function getSelectedAllergens() {
-  const checkboxes = document.querySelectorAll('input[name="allergy"]:checked');
-  const customAllergies = document.getElementById('otherAllergies')?.value
-    .split(',')
-    .map(s => s.trim())
-    .filter(s => s.length > 0) || [];
-
-  const selectedAllergies = Array.from(checkboxes).map(cb => cb.value);
-  return [...selectedAllergies, ...customAllergies];
-}
-
-/**
- * Save new recipe
- */
-async function saveRecipe(event) {
-  event.preventDefault();
-
-  // Get form values once
-  const dietaryTags = getSelectedDietaryTags();
-  const allergens = getSelectedAllergens();
-
-  const ingredientsText = document.getElementById('ingredients')?.value || '';
-  const ingredientsArray = ingredientsText
-    .split('\n')
-    .map(line => line.trim())
-    .filter(line => line.length > 0);
-
-  const stepsText = document.getElementById('steps')?.value || '';
-  const stepsArray = stepsText
-    .split('\n')
-    .map(line => line.trim())
-    .filter(line => line.length > 0);
-
-  const recipeData = {
-    name: document.getElementById('recipeName')?.value,
-    prep_time: parseInt(document.getElementById('prepTime')?.value),
-    cost: document.getElementById('cost')?.value,
-    difficulty: document.getElementById('difficulty')?.value,
-    ingredients: ingredientsArray,
-    prep_steps: stepsArray,
-    dietary_tags: dietaryTags,
-    allergens: allergens
-  };
-
-  console.log('Saving recipe with data:', recipeData);
-
-  // Validate required fields
-  if (!recipeData.name || !recipeData.prep_time || !ingredientsArray.length || !stepsArray.length) {
-    alert('Please fill in all required fields');
-    return;
-  }
-
-  try {
-    const response = await fetch(`${API_BASE_URL}/recipes`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${getToken()}`
-      },
-      body: JSON.stringify(recipeData)
-    });
-
-    const responseData = await response.json();
-    console.log('Server response:', responseData);
-
-    if (response.ok) {
-      alert('Recipe saved successfully!');
-      window.location.href = 'recipes.html';
-    } else {
-      alert('Error: ' + (responseData.message || 'Failed to save recipe'));
-    }
-  } catch (error) {
-    console.error('Error saving recipe:', error);
-    alert('Failed to save recipe. Please try again.');
-  }
-}
-
-/**
- * Populate edit form with recipe data
- */
-function populateEditForm(recipe) {
-  // Basic fields
-  setFieldValue('recipeName', recipe.name);
-  setFieldValue('prepTime', recipe.prep_time);
-  setFieldValue('cost', recipe.cost);
-  setFieldValue('difficulty', recipe.difficulty);
-
-  // Convert ingredients to text for editing
-  if (recipe.ingredients && Array.isArray(recipe.ingredients)) {
-    setFieldValue('ingredients', recipe.ingredients.join('\n'));
-  }
-
-  // Convert steps to text for editing
-  if (recipe.prep_steps && Array.isArray(recipe.prep_steps)) {
-    setFieldValue('steps', recipe.prep_steps.join('\n'));
-  }
-
-
-  // Dietary tags checkboxes
-  if (recipe.dietary_tags && Array.isArray(recipe.dietary_tags)) {
-    recipe.dietary_tags.forEach(tag => {
-      // Map database values back to form values
-      let formValue = tag;
-      if (tag === 'Gluten-Free') formValue = 'gluten-free';
-      if (tag === 'Dairy-Free') formValue = 'dairy-free';
-      if (tag === 'Nut-Free') formValue = 'nut-free';
-      if (tag === 'Vegan') formValue = 'vegan';
-      if (tag === 'Vegetarian') formValue = 'vegetarian';
-      if (tag === 'Halal') formValue = 'halal';
-
-      const checkbox = document.querySelector(`input[name="dietary"][value="${formValue}"]`);
-      if (checkbox) checkbox.checked = true;
-    });
-  }
-
-  // Allergens checkboxes
-  if (recipe.allergens && Array.isArray(recipe.allergens)) {
-    const customAllergies = [];
-
-    recipe.allergens.forEach(allergen => {
-      const checkbox = document.querySelector(`input[name="allergy"][value="${allergen.toLowerCase()}"]`);
-      if (checkbox) {
-        checkbox.checked = true;
-      } else {
-        customAllergies.push(allergen);
-      }
-    });
-
-    // Set custom allergies
-    if (customAllergies.length > 0) {
-      const otherAllergies = document.getElementById('otherAllergies');
-      if (otherAllergies) {
-        otherAllergies.value = customAllergies.join(', ');
-      }
-    }
-  }
-}
-
-/**
- * Helper to set field value
- */
-function setFieldValue(id, value) {
-  const field = document.getElementById(id);
-  if (field && value) {
-    field.value = value;
-  }
-}
-
-/**
- * Update recipe
- */
-async function updateRecipe(recipeId) {
-  // Get form values
-  const dietaryTags = getSelectedDietaryTags();
-  const allergens = getSelectedAllergens();
-
-  // Convert ingredients from text to array
-  const ingredientsText = document.getElementById('ingredients')?.value || '';
-  const ingredientsArray = ingredientsText
-    .split('\n')
-    .map(line => line.trim())
-    .filter(line => line.length > 0);
-
-  // Convert steps from text to array
-  const stepsText = document.getElementById('steps')?.value || '';
-  const stepsArray = stepsText
-    .split('\n')
-    .map(line => line.trim())
-    .filter(line => line.length > 0);
-
-  const recipeData = {
-    name: document.getElementById('recipeName')?.value,
-    prep_time: parseInt(document.getElementById('prepTime')?.value),
-    cost: document.getElementById('cost')?.value,
-    difficulty: document.getElementById('difficulty')?.value,
-    ingredients: ingredientsArray,
-    prep_steps: stepsArray,
-    dietary_tags: dietaryTags,
-    allergens: allergens
-  };
-
-  console.log('Updating recipe with data:', recipeData);
-
-  // Validate required fields
-  if (!recipeData.name || !recipeData.prep_time || !ingredientsArray.length || !stepsArray.length) {
-    alert('Please fill in all required fields');
-    return;
-  }
-
-  try {
-    const response = await fetch(`${API_BASE_URL}/recipes/${recipeId}`, {
-      method: 'PUT',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${getToken()}`
-      },
-      body: JSON.stringify(recipeData)
-    });
-
-    const responseData = await response.json();
-    console.log('Server response:', responseData);
-
-    if (response.ok) {
-      alert('Recipe updated successfully!');
-      window.location.href = 'recipes.html';
-    } else {
-      alert('Error: ' + (responseData.message || 'Failed to update recipe'));
-    }
-  } catch (error) {
-    console.error('Error updating recipe:', error);
-    alert('Failed to update recipe. Please try again.');
-  }
-}
-
-// MEAL PLANNER FUNCTIONS
-function setTodayDate() {
-    const today = new Date();
-    // Format as YYYY-MM-DD
-    const formattedDate = today.toISOString().split('T')[0];
-    
-    document.getElementById("weekPicker").value = formattedDate;
-
-    mealPlannerDate();
-}
-
-function parseLocalDate(dateString) {
-    const [year, month, day] = dateString.split('-').map(Number);
-    return new Date(year, month - 1, day); // months are 0-indexed
-}
-
-function mealPlannerDate() {
-    const dateContainer = document.getElementById("mealplanWeekDate");
-    dateContainer.innerHTML = "";
-
-    const spanMonth = document.createElement("span");
-    const spanWeek = document.createElement("span");
-
-    spanMonth.className = "week-month";
-    spanWeek.className = "week-range";
-
-    const inputValue = document.getElementById("weekPicker").value;
-
-    const selectedDate = inputValue ? parseLocalDate(inputValue) : new Date();
-
-    // Month
-    const monthFormatter = new Intl.DateTimeFormat("en-US", {
-        month: "long",
-        year: "numeric"
-    });
-    spanMonth.textContent = monthFormatter.format(selectedDate);
-    
-    // Week range
-    const startOfWeek = new Date(selectedDate);
-    const day = startOfWeek.getDay(); // 0 = Sunday
-    const diffToMonday = (day === 0 ? -6 : 1 - day); // adjust to Monday
-
-    startOfWeek.setDate(selectedDate.getDate() + diffToMonday);
-
-    const endOfWeek = new Date(startOfWeek);
-    endOfWeek.setDate(startOfWeek.getDate() + 6);
-
-    const shortFormatter = new Intl.DateTimeFormat("en-US", {
-        month: "short",
-        day: "numeric"
-    });
-    spanWeek.textContent = `${shortFormatter.format(startOfWeek)} - ${shortFormatter.format(endOfWeek)}`;
-    
-    dateContainer.appendChild(spanMonth);
-    dateContainer.appendChild(spanWeek);
-
-    weeklyMeals(selectedDate);
-}
-
-function mealPlannerWeek(input) {
-    const inputDate = document.getElementById("weekPicker");
-
-    let date = inputDate.value ? parseLocalDate(inputDate.value) : new Date();
-
-    const currentDay = date.getDay(); // 0 = Sunday
-    const diffToMonday = (currentDay === 0 ? -6 : 1 - currentDay);
-    date.setDate(date.getDate() + diffToMonday);
-
-    // Move 7 days
-    if(input == "next"){
-        date.setDate(date.getDate() + 7);
-    } else if (input == "prev"){
-        date.setDate(date.getDate() - 7);
-    }
-
-    // Update input value
-    inputDate.value = date.toISOString().split("T")[0];
-
-    mealPlannerDate();
-}
-
-console.log("weeklyMeals called");
-
-// Load meal plan for the week and populate the grid
-async function weeklyMeals(currentDate) {
-  const plannerGrid = document.getElementsByClassName("planner-grid")[0];
-  plannerGrid.innerHTML = "";
-
-  // create header corner
-  const corner = document.createElement("div");
-  corner.className = "grid-header-corner";
-  plannerGrid.appendChild(corner);
-
-  const currentDay = currentDate.getDay(); // 0 = Sunday
-  const diffToMonday = (currentDay === 0 ? -6 : 1 - currentDay);
-  const monday = new Date(currentDate);
-  monday.setDate(currentDate.getDate() + diffToMonday);
-
-  // header row
-  for (let i = 0; i < 7; i++) {
-    const day = document.createElement("div");
-
-    const dateForDay = new Date(monday);
-    dateForDay.setDate(monday.getDate() + i);
-
-    const today = new Date();
-    if (
-      dateForDay.getDate() === today.getDate() &&
-      dateForDay.getMonth() === today.getMonth() &&
-      dateForDay.getFullYear() === today.getFullYear()
-    ) {
-      day.className = "grid-day-header today";
-    } else {
-      day.className = "grid-day-header";
-    }
-
-    const nameDay = document.createElement("span");
-    const dateDay = document.createElement("span");
-    nameDay.className = "day-name";
-    dateDay.className = "day-date";
-
-    nameDay.textContent = dateForDay.toLocaleDateString("en-US", { weekday: "short" });
-    dateDay.textContent = dateForDay.toLocaleDateString("en-US", { day: "numeric" });
-
-    day.appendChild(nameDay);
-    day.appendChild(dateDay);
-
-    plannerGrid.appendChild(day);
-  }
-
-  // get meal plan for the week
-  const mondayStr = `${monday.getFullYear()}-${String(monday.getMonth() + 1).padStart(2, '0')}-${String(monday.getDate()).padStart(2, '0')}`;
-  console.log("mondayStr being requested:", mondayStr);
-
-  const mealPlanResponse = await fetch(`${API_BASE_URL}/mealPlan/week/${mondayStr}`, {
-    headers: {
-      'Authorization': `Bearer ${getToken()}`
-    }
-  });
-
-  console.log("mealPlanResponse status:", mealPlanResponse.status);
-
-  const mealPlan = await mealPlanResponse.json();
-  console.log("mealPlan data:", mealPlan)
-  
-  let mealPlanItems = [];
-
-  if (mealPlan) {
-    const itemsResponse = await fetch(`${API_BASE_URL}/mealPlan/${mealPlan.id}/items`, {
-      headers: {
-        'Authorization': `Bearer ${getToken()}`
-      }
-    });
-
-    console.log("itemsResponse status:", itemsResponse.status);
-
-    mealPlanItems = await itemsResponse.json();
-
-    console.log("mealPlanItems data:", mealPlanItems);
-  }
-
-  const MEAL_TYPES = [
-    { type: "BREAKFAST", icon: "🍳", label: "Breakfast" },
-    { type: "LUNCH", icon: "🥗", label: "Lunch" },
-    { type: "DINNER", icon: "🍽️", label: "Dinner" },
-    { type: "SNACK", icon: "🍎", label: "Snack" }
-  ];
-
-  const WEEKDAY_ENUM = ["SUNDAY", "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY"];
-  const startOfWeek = new Date(monday);
-
-  for (const meal of MEAL_TYPES) {
-    const mealTypeLabel = document.createElement("div");
-    mealTypeLabel.className = "grid-meal-label";
-
-    const icon = document.createElement("span");
-    icon.className = "meal-label-icon";
-    icon.textContent = meal.icon;
-
-    const text = document.createElement("span");
-    text.textContent = meal.label;
-
-    mealTypeLabel.appendChild(icon);
-    mealTypeLabel.appendChild(text);
-    plannerGrid.appendChild(mealTypeLabel);
-
-    for (let i = 0; i < 7; i++) {
-      const grid = document.createElement("div");
-      grid.className = "grid-cell";
-
-      const dateForDay = new Date(startOfWeek);
-      dateForDay.setDate(startOfWeek.getDate() + i);
-      const dayEnum = WEEKDAY_ENUM[dateForDay.getDay()];
-
-      const item = mealPlanItems.find(e =>
-        e.meal_type === meal.type &&
-        e.day_of_week === dayEnum
-      );
-
-    if (item) {
-        try {
-          // Fetch recipe details for the assigned meal
-          const recipe = item.recipe;
-
-          const cellFilled = document.createElement("div");
-          cellFilled.className = "cell-filled";
-
-          const recipeName = document.createElement("div");
-          recipeName.className = "cell-recipe-name";
-          recipeName.textContent = recipe.name;
-
-          const recipeMeta = document.createElement("div");
-          recipeMeta.className = "cell-recipe-meta";
-          recipeMeta.textContent = `⏱️ ${recipe.prep_time} min · ${recipe.difficulty}`;
-
-          const actions = document.createElement("div");
-          actions.className = "cell-filled-actions";
-
-          const editBtn = document.createElement("button");
-          editBtn.className = "cell-action-btn cell-btn-edit";
-          editBtn.type = "button";
-          editBtn.textContent = "✏️ Edit";
-          editBtn.addEventListener("click", () => {
-            saveSelectedMealSlot({
-              itemId: item.id,
-              mealPlanId: item.mealPlanId,
-              day_of_week: item.day_of_week,
-              meal_type: item.meal_type,
-              notes: item.notes || ""
-            });
-
-            window.location.href = "add-meal.html";
-          });
-
-          const removeBtn = document.createElement("button");
-          removeBtn.className = "cell-action-btn cell-btn-remove";
-          removeBtn.type = "button";
-          removeBtn.textContent = "✕";
-
-          // Add confirmation before deletion
-            removeBtn.addEventListener("click", async () => {
-            const confirmed = confirm(`Are you sure you want to remove "${recipe.name}" from your planner?`);
-            if (!confirmed) return;
-
-          // Store details of the deleted meal for potential undo functionality
-              lastDeletedMeal = {
-              mealPlanId: item.mealPlanId,
-              recipeId: item.recipeId,
-              day_of_week: item.day_of_week,
-              meal_type: item.meal_type,
-              notes: item.notes || ""
-            };
-
-            try {
-              const response = await fetch(`${API_BASE_URL}/mealPlan/items/${item.id}`, {
-                method: "DELETE",
-                headers: {
-                  "Authorization": `Bearer ${getToken()}`
-                }
-              });
-
-              const data = await response.json();
-
-              if (!response.ok) {
-                alert(data.message || "Failed to delete meal assignment.");
-                return;
-              }
-              // Refresh the meal plan grid after deletion
-              await weeklyMeals(currentDate);
-              showUndoBanner(); // Show undo option after deletion
-              
-            } catch (error) {
-              console.error("Error deleting meal assignment:", error);
-              alert("Something went wrong while deleting the meal.");
-            }
-          });
-
-          actions.appendChild(editBtn);
-          actions.appendChild(removeBtn);
-
-          cellFilled.appendChild(recipeName);
-          cellFilled.appendChild(recipeMeta);
-          cellFilled.appendChild(actions);
-
-          grid.appendChild(cellFilled);
-        } catch (error) {
-          console.error("Error loading recipe details:", error);
-        }
-      } else {
-        grid.className = "grid-cell empty";
-
-        const addBtn = document.createElement("button");
-        addBtn.type = "button";
-        addBtn.className = "cell-add-btn";
-        addBtn.innerHTML = `
-          <span class="add-icon">＋</span>
-          <span>Add ${meal.label}</span>
-        `;
-
-      addBtn.addEventListener("click", async () => {
-        let currentMealPlan = mealPlan;
-
-        if (!currentMealPlan) {
-          const weekStart = new Date(monday);
-          const weekEnd = new Date(monday);
-          weekEnd.setDate(weekEnd.getDate() + 6);
-
-          const result = await createMealPlanForWeek(
-            weekStart.toISOString(),
-            weekEnd.toISOString()
-          );
-
-          if (!result.success) {
-            alert(result.error || "Failed to create meal plan.");
-            return;
-          }
-
-          currentMealPlan = result.data;
-        }
-
-        saveSelectedMealSlot({
-          mealPlanId: currentMealPlan.id,
-          day_of_week: dayEnum,
-          meal_type: meal.type
-        });
-
-        window.location.href = "add-meal.html";
-      });
-
-        grid.appendChild(addBtn);
-      }
-
-      plannerGrid.appendChild(grid);
-    }
-  }
-}
-
-// Function to create a new meal plan item
-async function createMealPlanItem(mealPlanId, recipeId, day_of_week, meal_type, notes = "", allowDuplicate = false) {
+export async function createMealPlanItem(mealPlanId, recipeId, day_of_week, meal_type, notes = "", allowDuplicate = false) {
   try {
     if (!getToken()) {
       return { success: false, error: "You must be logged in." };
@@ -1853,1288 +335,45 @@ async function createMealPlanItem(mealPlanId, recipeId, day_of_week, meal_type, 
   }
 }
 
-async function updateMealPlanItem(itemId, recipeId, day_of_week, meal_type, notes = "", allowDuplicate = false) {
-  try {
-    if (!getToken()) {
-      return { success: false, error: "You must be logged in." };
-    }
-
-    if (!itemId || Number.isNaN(parseInt(itemId, 10))) {
-      return { success: false, error: "Invalid meal assignment ID." };
-    }
-
-    if (!recipeId || Number.isNaN(parseInt(recipeId, 10))) {
-      return { success: false, error: "Invalid recipe selected." };
-    }
-
-    if (!day_of_week || !meal_type) {
-      return { success: false, error: "Missing meal assignment details." };
-    }
-
-    const response = await fetch(`${API_BASE_URL}/mealPlan/items/${parseInt(itemId, 10)}`, {
-      method: 'PUT',
-      headers: {
-        'Authorization': `Bearer ${getToken()}`,
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        recipeId: parseInt(recipeId, 10),
-        day_of_week,
-        meal_type,
-        notes,
-        allowDuplicate
-      })
-    });
-
-    const data = await response.json();
-
-    if (!response.ok) {
-      return {
-        success: false,
-        error: data.message || "Failed to update meal assignment.",
-        code: data.code,
-        duplicates: data.duplicates
-      };
-    }
-
-    return { success: true, data };
-  } catch (error) {
-    console.error("Error updating meal plan item:", error);
-    return { success: false, error: "Something went wrong while updating the meal." };
-  }
-}
-// Helper function to format duplicate meal assignment locations for user confirmation
-function formatDuplicateMealAssignments(duplicates = []) {
-  if (!Array.isArray(duplicates) || duplicates.length === 0) {
-    return "another slot";
-  }
-  // Map the duplicate assignments to user-friendly strings and join them with commas
-  return duplicates
-    .map(({ day_of_week, meal_type }) => {
-      const prettyDay = day_of_week
-        ? day_of_week.charAt(0) + day_of_week.slice(1).toLowerCase()
-        : "Unknown day";
-      const prettyMeal = meal_type
-        ? meal_type.charAt(0) + meal_type.slice(1).toLowerCase()
-        : "Unknown meal";
-      return `${prettyDay} ${prettyMeal}`;
-    })
-    .join(", ");
-}
-
-async function loadRecipesForAddMealPage() {
-  try {
-    const response = await fetch(`${API_BASE_URL}/recipes`, {
-      headers: {
-        Authorization: `Bearer ${getToken()}`
-      }
-    });
-
-    const recipes = await response.json();
-    const container = document.getElementById("addMealRecipesContainer");
-
-    if (!container) return;
-
-    container.innerHTML = "";
-
-    recipes.forEach((recipe) => {
-      const card = document.createElement("a");
-      card.href = "#";
-      card.className = "recipe-card";
-      card.dataset.recipeId = recipe.id;
-
-      card.innerHTML = `
-        <div class="recipe-image">🍽️</div>
-        <div class="recipe-info">
-          <h4>${recipe.name}</h4>
-          <div class="recipe-details">
-            <span>⏱️ ${recipe.prep_time} min</span>
-            <span>${recipe.difficulty}</span>
-            <span>💰 ${recipe.cost}</span>
-          </div>
-        </div>
-      `;
-
-      container.appendChild(card);
-    });
-  } catch (error) {
-    console.error("Error loading recipes for add meal page:", error);
-  }
-}
-
-function validateSelectedMealSlot(slot) {
-  const validDays = [
-    "MONDAY",
-    "TUESDAY",
-    "WEDNESDAY",
-    "THURSDAY",
-    "FRIDAY",
-    "SATURDAY",
-    "SUNDAY"
-  ];
-
-  const validMealTypes = [
-    "BREAKFAST",
-    "LUNCH",
-    "DINNER",
-    "SNACK"
-  ];
-
-  if (!slot) {
-    return { valid: false, message: "No meal slot selected." };
-  }
-
-  if (!slot.mealPlanId || Number.isNaN(parseInt(slot.mealPlanId, 10))) {
-    return { valid: false, message: "Invalid meal plan selected." };
-  }
-
-  if (!slot.day_of_week || !validDays.includes(slot.day_of_week)) {
-    return { valid: false, message: "Invalid day selected." };
-  }
-
-  if (!slot.meal_type || !validMealTypes.includes(slot.meal_type)) {
-    return { valid: false, message: "Invalid meal type selected." };
-  }
-
-  return { valid: true };
-}
-
-// PERSONAL FEATURE FUNCTIONS
-async function showUserProfile() {
-  const container = document.getElementById("userProfileHero");
-  container.innerHTML = "";
-
-  const response = await fetch(`/api/profile`, {
-      method: 'GET',
-      headers: {
-        'Authorization': `Bearer ${getToken()}`,
-      },
-    });
-
-  const user = await response.json();
-  
-  const responseRecipes = await fetch(`/api/recipes`, {
-    method: 'GET',
-    headers: {
-          'Authorization': `Bearer ${getToken()}`
-        }
-  });
-  const userRecipes = await responseRecipes.json();
-  const publicRecipes = userRecipes.filter(e => e.is_private == false);
-  const privateRecipes = userRecipes.filter(e => e.is_private == true);
-
-  container.innerHTML = `
-    <div class="avatar-circle avatar-large">${user.firstName[0] + user.lastName[0]}</div>
-      <div class="user-hero-info">
-        <h1 class="user-display-name">${user.firstName + " " + user.lastName}</h1>
-        <p class="user-handle">@${user.email.substring(0, user.email.lastIndexOf('@'))}</p>
-        <div class="user-hero-stats">
-          <div class="hero-stat">
-              <span class="hero-stat-num">${publicRecipes.length}</span>
-              <span class="hero-stat-label">Public Recipes</span>
-          </div>
-          <div class="hero-stat">
-              <span class="hero-stat-num">${privateRecipes.length}</span>
-              <span class="hero-stat-label">Private Recipes</span>
-          </div>
-        </div>
-      </div>
-    </div>
-  `;
-}
-async function showPublicRecipes() {
-  const container = document.getElementById("tab-public");
-  container.innerHTML = "";
-
-  const responseRecipes = await fetch(`/api/recipes`, {
-    method: 'GET',
-    headers: {
-          'Authorization': `Bearer ${getToken()}`
-        }
-  });
-
-  const userRecipes = await responseRecipes.json();
-  const publicRecipes = userRecipes.filter(e => e.is_private == false);
-
-  for (const e of publicRecipes) {
-    //ratings
-    const responseRating = await fetch(`/api/averageRating/${e.id}`);
-    const avg = await responseRating.json();
-    const rounded = Math.round(avg);
-    // Build stars string
-    const stars = '★'.repeat(rounded) + '☆'.repeat(5 - rounded);
-
-    //number of ratings
-    const responseRatings = await fetch(`/api/recipeRatings/${e.id}`);
-    const ratings = await responseRatings.json();
-    let numberOfRatings = ratings.length;
-    if(!numberOfRatings){
-      numberOfRatings = 0;
-    }
-    //comments
-    const comments = ratings.filter(r => r.comment && r.comment.trim() !== "");
-    const commentsHTML = comments.map(r => `
-      <div class="comment-item">
-        <div class="comment-avatar">
-          ${r.user?.firstName ? r.user.firstName[0] : "?"}
-        </div>
-        <div class="comment-body">
-          <div class="comment-header">
-            <span class="comment-username">
-              @${r.user?.email ? r.user.email.substring(0, r.user.email.lastIndexOf('@')) : "unknown"}
-            </span>
-            <span class="comment-time">${new Date(r.updated_at).toLocaleDateString()}</span>
-          </div>
-          <p class="comment-text">${r.comment}</p>
-        </div>
-      </div>
-    `).join("");
-
-    //video
-    const firstVideo = Array.isArray(e.videos) && e.videos.length > 0
-      ? e.videos[0]
-      : null;
-
-    const videoLink = firstVideo?.videoUrl || null;
-
-    let videoHTML = videoLink
-      ? `
-        <a href="${videoLink}" target="_blank" class="video-link">
-          ▶️ Watch Video
-        </a>
-      `
-      : `
-        <p>No video available.</p>
-      `;
-
-    const recipeCard = document.createElement("div");
-    recipeCard.className = "recipe-card-full";
-
-    recipeCard.innerHTML = `
-        <details>
-          <summary class="recipe-card-top">
-            <span class="recipe-card-emoji">🍽️</span>
-            <div class="recipe-card-summary">
-                <h3 class="recipe-card-title">${e.name}</h3>
-                <div class="recipe-card-tags">
-                    <span class="tag tag-${e.difficulty.toLowerCase()}">${e.difficulty}</span>
-                    <span class="tag tag-cost">💰 ${e.cost}</span>
-                </div>
-                <div class="recipe-card-meta">
-                    <span>⏱️ ${e.prep_time} min</span>
-                      <span class="recipe-rating">
-                          <span class="stars-display">${stars}</span>
-                          <span class="rating-score">${avg.toFixed(1)}</span>
-                          <span class="rating-count">(${numberOfRatings} ratings)</span>
-                      </span>
-                </div>
-            </div>
-            <span class="recipe-card-chevron">▼</span>
-        </summary>
-        <div class="recipe-card-details">
-            <div class="recipe-video">
-              <h4>🎥 Video</h4>
-              ${videoHTML}
-            </div>
-            <div class="recipe-card-ingredients">
-                <strong>Ingredients:</strong><br>
-                ${e.ingredients.join(' · ')}
-            </div>
-            <div class="recipe-card-steps">
-                <strong>Steps:</strong>
-                <ol>
-                    ${e.prep_steps.map(step => `<li>${step}</li>`).join('')}
-                </ol>
-            </div>
-
-            <p class="rate-prompt">Your recipe's ratings</p>
-            <div class="recipe-rating">
-                <span class="stars-display">${stars}</span>
-                <span class="rating-score">${avg.toFixed(1)}</span>
-                <span class="rating-count">(${numberOfRatings} ratings)</span>
-            </div>
-
-            <div class="recipe-comments">
-              <h4 class="comments-heading">
-                💬 Comments <span class="comments-count">${comments.length}</span>
-              </h4>
-
-              <div class="comments-list">
-                ${commentsHTML || "<p>No comments yet.</p>"}
-              </div>
-            </div>
-
-            <div class="video-upload-form">
-                <label class="video-upload-label">
-                    🎥 Add a video or a link to a URL
-                    <input type="file" id="videoInput-${e.id}" name="video" accept="video/*">
-                </label>
-                <div class="video-url-section">
-                  <input type="text" id="urlInput-${e.id}" placeholder="Paste video URL (YouTube, etc.)">
-                </div>
-                <button type="button" class="btn-secondary" onclick="uploadVideo(${e.id})">Upload</button>
-            </div>
-
-            <div class="recipe-card-actions">
-                <button type="button" class="btn-make-public" onclick="changePrivacy(${e.id}, true)">🌍 Make Private</button>
-            </div>
-          </div>
-        </details>
-
-    `;
-    container.appendChild(recipeCard);
-  };
-}
-
-async function showPrivateRecipes() {
-  const container = document.getElementById("tab-private");
-  container.innerHTML = "";
-
-  const note = document.createElement("p");
-  note.className = "private-note";
-  note.textContent = "🔒 Only you can see these recipes. They are automatically added after creating a recipe."
-  container.appendChild(note);
-
-  const responseRecipes = await fetch(`/api/recipes`, {
-    method: 'GET',
-    headers: {
-          'Authorization': `Bearer ${getToken()}`
-        }
-  });
-
-  const userRecipes = await responseRecipes.json();
-  const privateRecipes = userRecipes.filter(e => e.is_private == true);
-
-  for (const e of privateRecipes){
-    const recipeCard = document.createElement("div");
-    recipeCard.className = "recipe-card-full";
-    
-    recipeCard.innerHTML = `
-        <details>
-          <summary class="recipe-card-top">
-            <span class="recipe-card-emoji">🍽️</span>
-            <div class="recipe-card-summary">
-                <h3 class="recipe-card-title">${e.name}</h3>
-                <div class="recipe-card-tags">
-                    <span class="tag tag-${e.difficulty.toLowerCase()}">${e.difficulty}</span>
-                    <span class="tag tag-cost">💰 ${e.cost}</span>
-                    <span class="tag tag-private">🔒 Private</span>
-                </div>
-                <div class="recipe-card-meta">
-                    <span>⏱️ ${e.prep_time} min</span>
-                </div>
-            </div>
-            <span class="recipe-card-chevron">▼</span>
-        </summary>
-        <div class="recipe-card-details">
-            <div class="recipe-card-ingredients">
-                <strong>Ingredients:</strong><br>
-                ${e.ingredients.join(' · ')}
-            </div>
-            <div class="recipe-card-steps">
-                <strong>Steps:</strong>
-                <ol>
-                    ${e.prep_steps.map(step => `<li>${step}</li>`).join('')}
-                </ol>
-            </div>
-            <div class="recipe-card-actions">
-                <a href="edit-recipe.html?id=${e.id}" class="btn-edit-recipe">✏️ Edit</a>
-                <button type="button" class="btn-delete-recipe" onclick="deleteRecipe(${e.id})">🗑️ Delete</button>
-                <button type="button" class="btn-make-public" onclick="changePrivacy(${e.id}, false)">🌍 Make Public</button>
-            </div>
-          </div>
-        </details>
-
-    `;
-    container.appendChild(recipeCard);
-  };
-}
-
-async function uploadVideo(id) {
-  const fileInput = document.getElementById(`videoInput-${id}`);
-  const urlInput = document.getElementById(`urlInput-${id}`);
-  const file = fileInput.files[0];
-  const url = urlInput.value.trim();
-  if (!file && !url) {
-    alert("Please select a video first.");
-    return;
-  }
-  const recipeId = id;
-
-  if (file) {
-    const sizeMB = file.size / (1024 * 1024);
-
-    if (sizeMB > MAX_SIZE_MB) {
-      alert("File is too large! Max is 50MB.");
-      return;
-    }
-  
-    const formData = new FormData();
-    formData.append("video", file);
-  
+// FETCHING DATA FUNCTIONS
+/**
+ * Fetch recipe data by ID
+ */
+export async function fetchRecipeData(recipeId) {
     try {
-      const response = await fetch(`/api/recipes/${recipeId}/video/upload`, {
-        method: "POST",
-        headers: {
-          Authorization: `Bearer ${getToken()}`
-        },
-        body: formData
-      });
-  
-      const data = await response.json();
-  
-      if (!response.ok) {
-        throw new Error(data.error || "Upload failed");
-      }
-  
-      alert("Video uploaded successfully!");
-      await showPublicRecipes();
-  
-    } catch (error) {
-      alert("Error uploading video");
-    }
-  }
-
-  if (url) {
-
-    try {
-      const parsed = new URL(url);
-      if (!(parsed.protocol === "http:" || parsed.protocol === "https:")) {
-        alert("Invalid url");
-        return;
-      }
-    } catch (err) {
-      alert("Invalid url");
-      return;
-    }
-
-    try {
-      const response = await fetch(`/api/recipes/${recipeId}/video/url`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          "Authorization": `Bearer ${getToken()}`
-        },
-        body: JSON.stringify({ videoUrl: url })
-      });
-  
-      const data = await response.json();
-  
-      if (!response.ok) {
-        throw new Error(data.error || "Upload failed");
-      }
-  
-      alert("Video URL saved!");
-      await showPublicRecipes();
-  
-    } catch (error) {
-      alert("Error uploading video");
-    }
-  }
-
-};
-
-//changes privacy
-async function changePrivacy(recipeId, privacy) {
-  try {
-    const response = await fetch(`/api/recipes/privacy/${recipeId}`, {
-      method: 'PUT',
-      headers: {
-        'Authorization': `Bearer ${getToken()}`,
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({ is_private: privacy })
-    });
-
-    if (!response.ok) {
-      throw new Error("Failed to update privacy");
-    }
-
-    showPublicRecipes();
-    showPrivateRecipes();
-    showUserProfile();
-  } catch (error) {
-    console.error("Error changing privacy:", error);
-    alert("Failed to update privacy.");
-  }
-}
-
-// Initialize add meal page 
-async function initAddMealPage() {
-  if (!isAuthenticated()) {
-    window.location.href = 'login-page.html';
-    return;
-  }
-
-  const slot = getSelectedMealSlot();
-  const slotValidation = validateSelectedMealSlot(slot);
-  const header = document.querySelector('.profile-header h1');
-
-  if (!slotValidation.valid) {
-    alert(slotValidation.message);
-    window.location.href = 'meal-planner.html';
-    return;
-  }
-
-  if (header && slot?.itemId) {
-    header.textContent = '✏️ Edit Meal';
-  }
-
-  await loadRecipesForAddMealPage();
-
-  const recipeCards = document.querySelectorAll('.recipe-card[data-recipe-id]');
-
-  if (!recipeCards.length) {
-    console.warn("No recipe cards found on add-meal page.");
-  }
-
-  recipeCards.forEach(card => {
-    const recipeId = card.dataset.recipeId;
-
-    if (!recipeId || Number.isNaN(parseInt(recipeId, 10))) {
-      return;
-    }
-
-    card.style.cursor = 'pointer';
-
-    card.addEventListener('click', async (e) => {
-      e.preventDefault();
-
-      card.style.pointerEvents = 'none';
-
-      let result = slot.itemId
-        ? await updateMealPlanItem(
-            slot.itemId,
-            recipeId,
-            slot.day_of_week,
-            slot.meal_type,
-            slot.notes || ""
-          )
-        : await createMealPlanItem(
-            slot.mealPlanId,
-            recipeId,
-            slot.day_of_week,
-            slot.meal_type
-          );
-      // If the API returns a duplicate recipe error, prompt the user for confirmation to allow duplicates
-      if (!result.success && result.code === 'DUPLICATE_RECIPE_IN_WEEK') {
-        const duplicateLocations = formatDuplicateMealAssignments(result.duplicates);
-        const confirmed = confirm(
-          `This recipe is already used this week in ${duplicateLocations}. Do you want to use it again?`
-        );
-        // If the user confirms, retry the API call with allowDuplicate set to true
-        if (confirmed) {
-          result = slot.itemId
-            ? await updateMealPlanItem(
-                slot.itemId,
-                recipeId,
-                slot.day_of_week,
-                slot.meal_type,
-                slot.notes || "", // for empty notes
-                true
-              )
-            : await createMealPlanItem(
-                slot.mealPlanId,
-                recipeId,
-                slot.day_of_week,
-                slot.meal_type,
-                "", // for empty notes
-                true
-              );
-        }
-      }
-
-      if (result.success) {
-        clearSelectedMealSlot();
-        alert(slot.itemId ? 'Meal updated successfully!' : 'Meal added successfully!');
-        window.location.href = 'meal-planner.html';
-      } else {
-        card.style.pointerEvents = 'auto';
-        alert(result.error || (slot.itemId ? 'Failed to update meal.' : 'Failed to add meal.'));
-      }
-    });
-  });
-
-  const logoutLink = document.querySelector('.nav-link.logout');
-  if (logoutLink) {
-    logoutLink.addEventListener('click', (e) => {
-      e.preventDefault();
-      clearToken();
-      window.location.href = 'login-page.html';
-    });
-  }
-}
-
-
-//Helper functions to manage selected meal slot in localStorage for editing
-function saveSelectedMealSlot(slotData) {
-  localStorage.setItem('selectedMealSlot', JSON.stringify(slotData));
-}
-
-function getSelectedMealSlot() {
-  const raw = localStorage.getItem('selectedMealSlot');
-  return raw ? JSON.parse(raw) : null;
-}
-
-function clearSelectedMealSlot() {
-  localStorage.removeItem('selectedMealSlot');
-}
-
-console.log("initMealPlannerPage called");
-
-// Undo functionality for meal deletion
-function showUndoBanner() {
-  let banner = document.getElementById("undoBanner");
-
-  if (!banner) {
-    banner = document.createElement("div");
-    banner.id = "undoBanner";
-    document.body.appendChild(banner);
-  }
-
-  banner.style.position = "fixed";
-  banner.style.bottom = "24px";
-  banner.style.right = "24px";
-  banner.style.background = "#ffffff";
-  banner.style.border = "1px solid #e5e7eb";
-  banner.style.borderLeft = "4px solid #16a34a";
-  banner.style.padding = "18px 22px";
-  banner.style.borderRadius = "12px";
-  banner.style.boxShadow = "0 8px 20px rgba(0,0,0,0.12)";
-  banner.style.zIndex = "9999";
-  banner.style.display = "flex";
-  banner.style.alignItems = "center";
-  banner.style.gap = "12px";
-  banner.style.minWidth = "300px";
-  banner.style.fontSize = "16px";
-  banner.style.transition = "opacity 0.3s ease, transform 0.3s ease";
-
-  banner.style.transform = "translateY(20px)";
-  banner.style.opacity = "0"; 
-
-  setTimeout(() => {
-  banner.style.opacity = "1";
-  banner.style.transform = "translateY(0)";
-}, 10);
-
-
-
-  banner.innerHTML = `
-  <span style="font-weight: 600; color: #111827;">
-    Meal removed — But not too late to undo
-  </span>
-  <button id="undoDeleteBtn" style="
-    background: #16a34a;
-    color: white;
-    border: none;
-    padding: 8px 16px;
-    border-radius: 8px;
-    cursor: pointer;
-    font-size: 14px;
-    font-weight: 500;
-  ">
-    Undo
-  </button>
-`;
-
-  const undoBtn = document.getElementById("undoDeleteBtn");
-
-  undoBtn.onmouseover = () => {
-    undoBtn.style.opacity = "0.9";
-  };
-
-  undoBtn.onmouseout = () => {
-    undoBtn.style.opacity = "1";
-  };
-
-  undoBtn.onclick = async () => {
-    if (!lastDeletedMeal) return;
-
-    const result = await createMealPlanItem(
-      lastDeletedMeal.mealPlanId,
-      lastDeletedMeal.recipeId,
-      lastDeletedMeal.day_of_week,
-      lastDeletedMeal.meal_type,
-      lastDeletedMeal.notes || ""
-    );
-
-    if (result.success) {
-      lastDeletedMeal = null;
-      banner.style.display = "none";
-      mealPlannerDate();
-    } else {
-      alert(result.error || "Failed to restore meal.");
-    }
-  };
-
-  if (undoTimeout) clearTimeout(undoTimeout);
-
-  undoTimeout = setTimeout(() => {
-    banner.style.opacity = "0";
-    banner.style.transform = "translateY(10px)";
-
-    setTimeout(() => {
-      banner.style.display = "none";
-      banner.style.opacity = "1";
-      banner.style.transform = "translateY(0)";
-      lastDeletedMeal = null;
-    }, 300);
-  }, 5000);
-
-  banner.onmouseenter = () => {
-  if (undoTimeout) clearTimeout(undoTimeout);
-};
-
-banner.onmouseleave = () => {
-  undoTimeout = setTimeout(() => {
-    banner.style.opacity = "0";
-    banner.style.transform = "translateY(10px)";
-
-    setTimeout(() => {
-      banner.style.display = "none";
-      lastDeletedMeal = null;
-    }, 300);
-  }, 3000);
-};
-
-}
-
-// Initialize meal planner page 
-function initMealPlannerPage() {
-  if (!isAuthenticated()) {
-    window.location.href = 'login-page.html';
-    return;
-  }
-
-  const weekPicker = document.getElementById('weekPicker');
-  if (weekPicker) {
-    setTodayDate();
-
-    weekPicker.addEventListener('change', () => {
-      mealPlannerDate();
-    });
-  }
-
-  const weekButtons = document.querySelectorAll('.week-nav-btn');
-  if (weekButtons.length >= 2) {
-    weekButtons[0].addEventListener('click', () => mealPlannerWeek('prev'));
-    weekButtons[1].addEventListener('click', () => mealPlannerWeek('next'));
-  }
-
-  const logoutLink = document.querySelector('.nav-link.logout');
-  if (logoutLink) {
-    logoutLink.addEventListener('click', (e) => {
-      e.preventDefault();
-      clearToken();
-      window.location.href = 'login-page.html';
-    });
-  }
-}
-
-// PAGE INITIALIZATION FUNCTIONS
-
-/**
- * Initialize login page
- */
-function initLoginPage() {
-  const loginForm = document.querySelector('form');
-  if (!loginForm) return;
-
-  loginForm.addEventListener('submit', async (e) => {
-    e.preventDefault();
-    clearAllErrors();
-
-    // Get form values
-    const email = document.querySelector('input[type="email"]').value.trim();
-    const password = document.querySelector('input[type="password"]').value;
-
-    // Validate inputs
-    if (!email) {
-      alert('Please enter your email address');
-      return;
-    }
-
-    if (!password) {
-      alert('Please enter your password');
-      return;
-    }
-
-    // Get submit button
-    const submitButton = loginForm.querySelector('button[type="submit"]');
-    setButtonLoading(submitButton, true);
-
-    // Attempt login
-    const result = await loginUser(email, password);
-
-    setButtonLoading(submitButton, false);
-
-    if (result.success) {
-      // Save token
-      saveToken(result.data.token);
-
-      // Redirect to home page
-      window.location.href = 'home-page.html';
-    } else {
-      // Show error
-      alert(result.error || 'Login failed. Please check your credentials and try again.');
-    }
-  });
-}
-
-/**
- * Initialize signup page
- */
-function initSignupPage() {
-  const signupForm = document.getElementById('signupform');
-  if (!signupForm) return;
-
-  // Password toggle functionality
-  const toggleButtons = document.querySelectorAll('.toggle-password');
-  toggleButtons.forEach(button => {
-    button.addEventListener('click', () => {
-      const targetId = button.dataset.target;
-      const input = document.getElementById(targetId);
-
-      if (input.type === 'password') {
-        input.type = 'text';
-        button.textContent = '👁️';
-      } else {
-        input.type = 'password';
-        button.textContent = '👁';
-      }
-    });
-  });
-
-  // Form submission
-  signupForm.addEventListener('submit', async (e) => {
-    e.preventDefault();
-    clearAllErrors();
-
-    // Get form values
-    const firstName = document.getElementById('firstName').value.trim();
-    const lastName = document.getElementById('lastName').value.trim();
-    const email = document.getElementById('email').value.trim();
-    const password = document.getElementById('password').value;
-    const confirmPassword = document.getElementById('confirm_password').value;
-
-    let hasErrors = false;
-
-    // Validate first name
-    if (!firstName) {
-      showError('firstNameError', 'First name is required');
-      hasErrors = true;
-    }
-
-    // Validate last name
-    if (!lastName) {
-      showError('lastNameError', 'Last name is required');
-      hasErrors = true;
-    }
-
-    // Validate email
-    if (!email) {
-      showError('emailError', 'Email is required');
-      hasErrors = true;
-    } else if (!validateEmail(email)) {
-      showError('emailError', 'Please enter a valid email address');
-      hasErrors = true;
-    }
-
-    // Validate password
-    const passwordValidation = validatePassword(password);
-    if (!passwordValidation.valid) {
-      showError('passwordError', passwordValidation.message);
-      hasErrors = true;
-    }
-
-    // Validate password confirmation
-    if (password !== confirmPassword) {
-      showError('confirmError', 'Passwords do not match');
-      hasErrors = true;
-    }
-
-    if (hasErrors) return;
-
-    // Get submit button
-    const submitButton = signupForm.querySelector('button[type="submit"]');
-    setButtonLoading(submitButton, true);
-
-    // Attempt registration
-    const result = await registerUser(firstName, lastName, email, password);
-
-    setButtonLoading(submitButton, false);
-
-    if (result.success) {
-      // Show success message
-      alert('Registration successful! Please log in.');
-
-      // Redirect to login page
-      window.location.href = 'login-page.html';
-    } else {
-      // Show error
-      if (result.error.includes('Email')) {
-        showError('emailError', result.error);
-      } else {
-        alert(result.error || 'Registration failed. Please try again.');
-      }
-    }
-  });
-}
-
-/**
- * Initialize edit profile page
- */
-function initEditProfilePage() {
-  // Check authentication
-  if (!isAuthenticated()) {
-    window.location.href = 'login-page.html';
-    return;
-  }
-
-  // Load user profile and populate form
-  loadUserProfileForEdit();
-
-  // Setup logout functionality
-  const logoutLink = document.querySelector('.nav-link.logout');
-  if (logoutLink) {
-    logoutLink.addEventListener('click', (e) => {
-      e.preventDefault();
-      clearToken();
-      window.location.href = 'login-page.html';
-    });
-  }
-
-  // Setup form submission
-  const profileForm = document.getElementById('profileForm');
-  if (profileForm) {
-    profileForm.addEventListener('submit', async (e) => {
-      e.preventDefault();
-      console.log('Form submitted!')
-      clearAllErrors();
-
-      // Collect form data
-      const formData = {
-        firstName: document.getElementById('firstName')?.value.trim() || '',
-        lastName: document.getElementById('lastName')?.value.trim() || '',
-        email: document.getElementById('email')?.value.trim() || '',
-        age: document.getElementById('editAge')?.value ? parseInt(document.getElementById('editAge').value) : null,
-        weight: document.getElementById('editWeight')?.value ? parseFloat(document.getElementById('editWeight').value) : null,
-        height: document.getElementById('editHeight')?.value ? parseFloat(document.getElementById('editHeight').value) : null,
-        dietPreferences: [],
-        allergies: [],
-        cookingSkill: document.getElementById('cookingSkill')?.value || '',
-        mealPrepTime: document.getElementById('mealPrepTime')?.value || '',
-        budgetRange: document.getElementById('budgetRange')?.value || '',
-      };
-
-      console.log('Form data collected:', formData);
-
-      // Get selected diet preferences
-      const dietCheckboxes = document.querySelectorAll('input[name="diet"]:checked');
-      dietCheckboxes.forEach(cb => formData.dietPreferences.push(cb.value));
-
-      // Get selected allergies
-      const allergyCheckboxes = document.querySelectorAll('input[name="allergy"]:checked');
-      const extraAllergyText = document.getElementById('otherAllergies')?.value.split(',').map(s => s.trim()).filter(s => s.length > 0);
-      allergyCheckboxes.forEach(cb => formData.allergies.push(cb.value));
-      if (extraAllergyText) {
-        extraAllergyText.forEach(element => formData.allergies.push(element));
-      }
-
-      // Validate
-      let hasErrors = false;
-
-      if (!formData.firstName || !formData.lastName) {
-        showError('nameError', 'Name is required');
-        hasErrors = true;
-      }
-
-      if (!formData.email) {
-        showError('emailError', 'Email is required');
-        hasErrors = true;
-      } else if (!validateEmail(formData.email)) {
-        showError('emailError', 'Please enter a valid email address');
-        hasErrors = true;
-      }
-
-      if (hasErrors) return;
-
-      // Get submit button
-      const submitButton = profileForm.querySelector('button[type="submit"]');
-      setButtonLoading(submitButton, true);
-
-      // Update profile
-      const result = await updateUserProfile(formData);
-
-      setButtonLoading(submitButton, false);
-
-      if (result.success) {
-        alert('Profile updated successfully!');
-        window.location.href = 'home-page.html';
-      } else {
-        alert(result.error || 'Failed to update profile. Please try again.');
-      }
-    });
-  }
-}
-
-/**
- * Initialize recipes page
- */
-function initRecipesPage() {
-  console.log('🍽️ INIT_RECIPES_PAGE CALLED');
-
-  // Check authentication
-  if (!isAuthenticated()) {
-    window.location.href = 'login-page.html';
-    return;
-  }
-
-  // Load My Recipes
-  loadMyRecipes();
-
-  // Load General Recipes
-  //loadGeneralRecipes();
-
-  // Load Explore Recipes
-  loadExploreRecipes();
-
-  // Setup logout functionality
-  const logoutLink = document.querySelector('.nav-link.logout');
-  if (logoutLink) {
-    logoutLink.addEventListener('click', (e) => {
-      e.preventDefault();
-      clearToken();
-      window.location.href = 'login-page.html';
-    });
-  }
-
-  // Setup search functionality
-  const searchBtn = document.querySelector('.search-btn');
-  const searchInput = document.getElementById('searchInput');
-
-  if (searchBtn && searchInput) {
-    searchBtn.addEventListener('click', () => {
-      console.log('Search button clicked');
-      const searchTerm = searchInput.value.trim();
-      filterRecipesBySearch(searchTerm);
-    });
-
-    searchInput.addEventListener('keyup', (e) => {
-      if (e.key === 'Enter') {
-        console.log('Enter key pressed');
-        const searchTerm = searchInput.value.trim();
-        filterRecipesBySearch(searchTerm);
-      }
-    });
-  } else {
-    console.log('Search button or input not found');
-  }
-
-  // Setup Apply Filters button
-  const applyFiltersBtn = document.getElementById('applyFiltersBtn');
-  console.log('Apply Filters button found:', applyFiltersBtn);
-
-  if (applyFiltersBtn) {
-    applyFiltersBtn.addEventListener('click', () => {
-      console.log('Apply Filters button clicked');
-      filterRecipes();
-    });
-  } else {
-    console.log('Apply Filters button NOT found - check if ID is correct in HTML');
-  }
-
-  // Setup Clear Filters button
-  const clearFiltersBtn = document.getElementById('clearFiltersBtn');
-    if (clearFiltersBtn) {
-    clearFiltersBtn.addEventListener('click', async () => {
-      console.log('Clear Filters button clicked');
-      document.querySelectorAll('.filter-bar input[type="checkbox"]').forEach(cb => {
-        cb.checked = false;
-      });
-      await displayFilteredRecipes(allRecipes);
-    });
-  }
-}
-
-/**
- * Initialize create recipe page
- */
-function initCreateRecipePage() {
-  console.log('📝 INIT_CREATE_RECIPE PAGE');
-
-  // Check authentication
-  if (!isAuthenticated()) {
-    window.location.href = 'login-page.html';
-    return;
-  }
-
-  const form = document.getElementById('createRecipeForm');
-  if (!form) return;
-
-  // Override the default form submission
-  form.addEventListener('submit', saveRecipe);
-
-  // Setup logout
-  const logoutLink = document.querySelector('.nav-link.logout');
-  if (logoutLink) {
-    logoutLink.addEventListener('click', (e) => {
-      e.preventDefault();
-      clearToken();
-      window.location.href = 'login-page.html';
-    });
-  }
-}
-
-/**
- * Initialize edit recipe page
- */
-async function initEditRecipePage() {
-  console.log('✏️ INIT_EDIT_RECIPE PAGE');
-
-  // Check authentication
-  if (!isAuthenticated()) {
-    window.location.href = 'login-page.html';
-    return;
-  }
-
-  // Get recipe ID from URL
-  const urlParams = new URLSearchParams(window.location.search);
-  const recipeId = urlParams.get('id');
-
-  if (!recipeId) {
-    alert('No recipe ID specified');
-    window.location.href = 'recipes.html';
-    return;
-  }
-
-  // Fetch recipe data
-  try {
     const response = await fetch(`${API_BASE_URL}/recipes/${recipeId}`, {
+      method: 'GET',  
       headers: {
         'Authorization': `Bearer ${getToken()}`
-      }
+        }
     });
 
     if (!response.ok) {
-      throw new Error('Failed to fetch recipe');
+        throw new Error('Failed to fetch recipe');
     }
 
     const recipe = await response.json();
 
-    // Populate form fields
-    populateEditForm(recipe);
+    return recipe;
+    } catch (error) {
+        console.error('Error fetching recipe:', error);
+    }
+}
+/**
+ * Fetch all recipes
+ */
+export async function fetchAllRecipes() {
+  try {
+    const response = await fetch(`${API_BASE_URL}/recipes`, {
+      method: 'GET',  
+      headers: {
+        'Authorization': `Bearer ${getToken()}`
+        }
+    });
 
+    const recipes = await response.json();
+    return recipes;
   } catch (error) {
-    console.error('Error fetching recipe:', error);
-    alert('Error loading recipe');
-    window.location.href = 'recipes.html';
-  }
-
-  // Setup form submission
-  const form = document.getElementById('editRecipeForm');
-  if (form) {
-    form.addEventListener('submit', async (e) => {
-      e.preventDefault();
-      await updateRecipe(recipeId);
-    });
-  }
-
-  // Setup logout
-  const logoutLink = document.querySelector('.nav-link.logout');
-  if (logoutLink) {
-    logoutLink.addEventListener('click', (e) => {
-      e.preventDefault();
-      clearToken();
-      window.location.href = 'login-page.html';
-    });
+    console.error('Error fetching all recipes:', error);
   }
 }
-
-function initUserProfile() {
-  // Check authentication
-  if (!isAuthenticated()) {
-    window.location.href = 'login-page.html';
-    return;
-  }
-
-  showUserProfile()
-  showPublicRecipes();
-  showPrivateRecipes();
-
-  // Setup logout functionality
-  const logoutLink = document.querySelector('.nav-link.logout');
-  if (logoutLink) {
-    logoutLink.addEventListener('click', (e) => {
-      e.preventDefault();
-      clearToken();
-      window.location.href = 'login-page.html';
-    });
-  }
-}
-
-
-// AUTO-INITIALIZATION ON PAGE LOAD
-document.addEventListener('DOMContentLoaded', () => {
-  // Detect which page we're on and initialize accordingly
-  const path = window.location.pathname;
-  const page = path.substring(path.lastIndexOf('/') + 1);
-
-  console.log('Current page:', page);
-  console.log('Full path:', path);
-
-  switch (page) {
-    case 'login-page.html':
-    case 'login':
-    case '':
-      getInitials();
-      initLoginPage();
-      break;
-
-    case 'sign-up-page.html':
-    case 'signup':
-      getInitials();
-      initSignupPage();
-      break;
-
-    case 'edit-profile.html':
-    case 'profile':
-      getInitials();
-      initEditProfilePage();
-      break;
-
-    case 'recipes.html':
-      getInitials();
-      initRecipesPage();
-      break;
-
-    case 'create-recipe.html':
-      getInitials();
-      initCreateRecipePage();
-      break;
-
-    case 'edit-recipe.html':
-      getInitials();
-      initEditRecipePage();
-      break;
-
-    case 'meal-planner.html':
-      getInitials();
-      initMealPlannerPage();
-      break;
-
-    case 'add-meal.html':
-      getInitials();
-      initAddMealPage();
-      break;
-    
-    case 'user-profile.html':
-      getInitials();
-      initUserProfile();
-      break;
-
-    default:
-      console.log('Page not recognized for auto-initialization');
-  }
-  
-});
