+++ public/sign-up-page.html
@@ -6,8 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Meal Planner Sign Up</title>
     <link rel="stylesheet" href="styles.css">
-     <script type="module" src="script.js"></script>
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600&display=swap" rel="stylesheet">
+    <script type="module" src="script-sign-up-page.js"></script>
 </head>
 
 <body>
