+++ public/user-profile.html
@@ -12,7 +12,7 @@
     <link rel="stylesheet" href="user-profile-styles.css">
     <link rel="stylesheet" href="add-recipe-styles.css">
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&display=swap" rel="stylesheet">
-     <script type="module" src="script.js" defer></script>
+    <script type="module" src="script-user-profile.js" defer></script>
 </head>
 
 <body>
